requirements.txt
opencv-python==4.5.5.64
Pillow==9.0.1
numpy==1.22.3